Working with NewApplication form

+++ app/src/main/java/com/s/samsungitschool/recom00/fragments/NewAppFragmentActivity.java
@@ -277,7 +277,7 @@ public class NewAppFragmentActivity extends Fragment {
                         loadSuccessfully = false;
 
                     } else {
-                        pointId = (long) Long.valueOf(addPointServerAns);
+                        pointId = Long.valueOf(addPointServerAns);
 
                         loadSuccessfully = true;
                     }
